refactor: reorganize app packages
