+++ app/adapters/__init__.py
@@ -1,0 +1,3 @@
+from . import hh, avito, amochats, amo_client
+
+__all__ = ["hh", "avito", "amochats", "amo_client"]
