+++ tests/test_worker_handlers.py
@@ -4,7 +4,7 @@ import sys
 import pytest
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from app import worker_rmq
+from app.services import worker_rmq
 
 
 @pytest.mark.parametrize(
